[storage] Move test_truncation out of db_debugger into regular test path
The `test_truncation` proptest was inside the `db_debugger/truncate` module,
which is gated behind `#[cfg(feature = "db-debugger")]`. Since CI never enables
that feature for tests, the test never actually ran.

Extract the core truncation logic from `Cmd::run()` into a shared
`run_truncation()` helper in `utils/truncation_helper.rs`, gated with
`#[cfg(any(test, feature = "db-debugger"))]`. Move the test to
`db/aptosdb_test.rs` where it calls the helper directly and will be picked up
by normal `cargo test` runs. Widen the feature gate on
`catch_up_state_merkle_db` and `get_usage_before_or_at` to also include `test`.
